Moving filters to Scopes namespace.
